Some improvements.
EntityService should specify every EntityServiceImpl, that means every EntityServiceImpl has to implement EntityService and work out the interface-methods (with @Override).

EntityRepository should be needed by EntityServiceImpl. And if something from the EntityServiceImpl is needed in another class, EntityService should be used and Spring will take the EntityServiceImpl which is active (by Profile etc.) and the EntityServiceImpl cares for fetching the data from EntityRepository and integrate business logic.

But with UserService from jhipster, there is no UserServiceImpl, so if UserRepository is needed, UserService should be used - as far as I can understand right now...